% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtions.R
\name{filter_tourism_data}
\alias{filter_tourism_data}
\title{Filter Tourism GDP Data by Country and Year Range}
\usage{
filter_tourism_data(data, country, year_range)
}
\arguments{
\item{data}{The tourism dataset.}

\item{country}{Character. The name of the country to filter by.}

\item{year_range}{Numeric vector of length 2. The start and end year for filtering.}
}
\value{
A data frame containing the filtered tourism data.
}
\description{
This function filters the tourism GDP dataset based on a specified country and year range. It is useful for extracting a subset of the data for detailed analysis or visualization.
}
