% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtions.R
\name{plot_tourism_gdp}
\alias{plot_tourism_gdp}
\title{Plot Tourism GDP Contribution Over Time}
\usage{
plot_tourism_gdp(data)
}
\arguments{
\item{data}{A data frame containing the filtered tourism GDP data.}
}
\value{
A ggplot2 object that displays the tourism GDP contribution over time.
}
\description{
This function generates a plot of tourism's contribution to GDP for a given country over time. Users can specify the data to be plotted, allowing for customized visualizations.
}
\examples{
data(tourism)
australia_data <- filter_tourism_data(tourism, "Australia", c(2008, 2020))
plot_tourism_gdp(australia_data)
}
