% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtions.R
\name{run_app}
\alias{run_app}
\title{Launch the EcoTourImpact Shiny App}
\usage{
run_app()
}
\value{
A Shiny app interface allowing users to interactively explore tourism GDP data.
}
\description{
The `runMyShinyApp` function launches an interactive Shiny app that enables users to explore the tourism GDP contribution across various countries and years. Users can filter by country, adjust the year range, choose between different plot types, and view summary statistics.
}
